Example repo
Problem
-------
In the current state bzl-gen-build is not really usable
outside of our own usage since it's relying on custom scripts and macros.
This is often necessitated by our tendency to use old version of Python etc.

Solution
--------
This adds an example repository to reimplement some of the
scripts and configurations using open source rules_python etc.