Fix deps expansion
Problem
-------
secondary supports `deps` to be suffixed with something like `__py`,
but it currently doesn't take in account of the label name abbrebiation
like `a/b` (`a/b:b`), creating deps to `a/b__py`, which won't exist.

Solution
--------
This expands `a/b` into `a/b:b` when used for `deps`.